Add notification resource (NOTE: create is currenly broken).
This currently does not work as the server returns unexpecetd values
when creating a notification.  Specifically, the created_at value is
a float insead of an integer, and the environments are a map of
UUID->name instead of an array of maps of UUID->name.  Because of those
issues this resource is largely untested and will need to be revisied
after they are resolved.