Tweak filters a bit.
"values" is now an array again, though it can only accept a single
value.  This is for future-proofing so we can add support for multiple
values at a later date with server-side filtering.

Rules don't (yet) support server-side filtering, so the client-side
filtering has been tweaked to match the server-side behavior of the
other types; instead of using glob, we do a case-insensitive substring
match.