Use glob-style matching in combination with server-side filtering.
This brings back the glob syntax in filters, but still makes use of
server-side filtering.  The idea is basically that we split the glob
on *, then use the longest substring to pass to the server for
filtering.  This gives us a superset of the results we want, so locally
we apply the glob to build the final result set.